Update the timing for sending the emails: service holder state.
